Added support to WashTower
Added support to WashTower (Washer and Dryer in one stack device). Using the device name as the driver name to fix issue were the WashTower - Washer has the alias "Washer" but the WashTower - Dryer has the 'wrong' alias "WashTower". Improved debug to show if a device is supported by the app.